Fix bug with text appearing below the input
Instead of deleting the [for="password"]::after css rule from style
sheet, have it altered to an empty string.

If you delete the css rule, the bug is that if you press fast two keys
that are the same in password and confirm password inputs, it deletes
the cssRule[0], two times! So, it would delete the rule we wanted AND
the font-family, and this would continue, until the style sheet wouldn't
have any rules left.